requirements.txt
pandas
pysqlite3-binary
secure-smtplib
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
openai==1.16.2
chroma-hnswlib==0.7.5
chromadb==0.5.4
streamlit==1.38.0
